Move build system spec into own file & expand coverage
Asserts against the file output rather than its presence to ensure it
includes what's needed.
