+++ spec/fixtures/site/content/index.html.erb
@@ -1,3 +1,2 @@
-<% "Hello, world!" %>
-
+<% @title = "Hello, world!" %>
 <h1>Hello from Soy</h1>
