+++ spec/fixtures/site/views/layout.html.erb
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <title><%= "Soy Sample Site" %></title>
+    <title><%= @title %></title>
   </head>
   <body>
     <%= yield %>
