+++ spec/system/cli_spec.rb
@@ -3,10 +3,6 @@
 require "spec_helper"
 
 RSpec.describe "CLI" do
-  def run_cmd(cmd)
-    `ruby -Ilib ./exe/soy #{cmd}`
-  end
-
   describe "soy" do
     it "outputs help" do
       output = run_cmd(nil)
@@ -19,25 +15,6 @@ RSpec.describe "CLI" do
       output = run_cmd("bogus")
       expect(output).to match(/`bogus` is not a command/)
       expect(output).to match(/Run `soy help` for list of commands/)
-    end
-  end
-
-  describe "soy build" do
-    let(:fixture_dir) { "spec/fixtures/site" }
-    let(:build_dir) { "#{fixture_dir}/build" }
-
-    before { FileUtils.rm_rf(build_dir) }
-
-    it "generates the site" do
-      output = run_cmd("build #{fixture_dir}")
-      expect(output).to match(/Building site/)
-      expect(output).to match(/built in \d.\d+ seconds/)
-      expect(File.exist?("#{build_dir}/index.html")).to be(true)
-    end
-
-    it "supports the b alias" do
-      output = run_cmd("b #{fixture_dir}")
-      expect(output).to match(/Building site/)
     end
   end
